requirements.txt
djangorestframework
django
Pillow
djangorestframework-simplejwt
requests
